# Kraken: Improve /line_reports perf