--- a/source/type/type_interfaces.cpp
+++ b/source/type/type_interfaces.cpp
@@ -91,11 +91,11 @@ bool HasMessages::has_applicable_message(const boost::posix_time::ptime& current
         if (!impact) {
             continue;  // pointer might still have become invalid
         }
-        if (line && impact->is_only_line_section() && !impact->is_line_section_of(*line)) {
-            continue;
-        }
-        if (line && impact->is_only_rail_section() && !impact->is_rail_section_of(*line)) {
-            continue;
+        if (line) {
+            if ((impact->is_only_line_section() && !impact->is_line_section_of(*line))
+                || (impact->is_only_rail_section() && !impact->is_rail_section_of(*line))) {
+                continue;
+            }
         }
         if (impact->is_valid(current_time, action_period)) {
             // if filter empty == no filter